(Chore): Optimise code for dataset validation, refactor variable and module names for readability [ss3722]
