+++ aignostic/tests/model_api/model/mock.py
@@ -1,5 +1,5 @@
 from fastapi import FastAPI
-from aignostic.pydantic_models.models import DataSet
+from aignostic.pydantic_models.data_models import DataSet
 
 app = FastAPI()
