+++ aignostic/tests/model_api/model/scikit_mock.py
@@ -2,7 +2,7 @@ from fastapi import FastAPI
 import numpy as np
 from sklearn.pipeline import Pipeline
 import pickle
-from aignostic.pydantic_models.models import DataSet
+from aignostic.pydantic_models.data_models import DataSet
 
 app = FastAPI()
