+++ aignostic/tests/model_api/unit/test_model_api.py
@@ -1,10 +1,10 @@
 from fastapi.testclient import TestClient
-from tests.modelAPI.model.scikit_mock import app as scikit_app
-from tests.modelAPI.model.mock import app as mock_app
+from tests.model_api.model.scikit_mock import app as scikit_app
+from tests.model_api.model.mock import app as mock_app
 import pandas as pd
 from folktables import ACSDataSource, ACSEmployment
 import pickle
-from tests.modelAPI.model.huggingface_binclassifier import app as huggingface_app
+from tests.model_api.model.huggingface_binclassifier import app as huggingface_app
 
 
 client_huggingface = TestClient(huggingface_app)
