+++ aignostic/tests/model_api/unit/test_pydantic_conversions.py
@@ -1,6 +1,6 @@
-import aignostic.pydantic_models.models as models
+import aignostic.pydantic_models.data_models as data_models
 import numpy as np
-from tests.modelAPI.model.mock import app as mock_app
+from tests.model_api.model.mock import app as mock_app
 from fastapi.testclient import TestClient
 import pandas as pd
 
@@ -12,7 +12,7 @@ def test_conversion_from_numpy():
     Test conversion from numpy array to pydantic model
     """
     arr = np.array([[1, 2, 3], [4, 5, 6]]).astype(float)
-    data = models.arr_to_JSON(arr)
+    data = data_models.arr_to_JSON(arr)
     data["column_names"] = ["a", "b", "c"]
     response = mock_app.post("/predict", json=data)
     assert response.status_code == 200, response.text
@@ -24,7 +24,7 @@ def test_conversion_from_pandas():
     """
     df = np.array([[1, 2, 3], [4, 5, 6]]).astype(float)
     df = pd.DataFrame(df, columns=["a", "b", "c"])
-    data = models.df_to_JSON(df)
+    data = data_models.df_to_JSON(df)
     response = mock_app.post("/predict", json=data)
     assert response.status_code == 200, response.text
 
@@ -33,6 +33,6 @@ def test_conversion_from_csv():
     """
     Test conversion from csv to pydantic model
     """
-    data = models.csv_to_JSON("tests/modelAPI/unit/csv_testfile.csv")
+    data = data_models.csv_to_JSON("tests/model_api/unit/csv_testfile.csv")
     response = mock_app.post("/predict", json=data)
     assert response.status_code == 200, response.text
